backend/pyproject.toml
[project]
name = "deskhelp-bot"
version = "0.1.0"
description = "A LiveKit-powered desk help support bot"
dependencies = [
    "livekit-agents[cartesia,deepgram,openai,silero,turn-detector]~=1.2",
    "python-dotenv",
    "sqlalchemy",
    "fastapi",
    "uvicorn",
    "livekit-plugins-noise-cancellation~=0.2",
    "livekit-api",
]
requires-python = ">=3.9"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
